refactor: centralize warm cache edge restoration
Amp-Thread-ID: https://ampcode.com/threads/T-019edbec-b1c2-74c0-9afb-ab2dffcbc96e
